requirements.txt
asgiref==3.3.1
boto3==1.17.3
botocore==1.20.3
Django==3.1.5
django-js-asset==1.2.2
django-storages==1.11.1
django-tinymce4-lite==1.8.0
jmespath==0.10.0
jsmin==2.2.2
Pillow==8.1.0
python-dateutil==2.8.1
python-decouple==3.4
python-http-client==3.3.1
pytz==2020.5
s3transfer==0.3.4
sendgrid==6.5.0
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
urllib3==1.26.3
